Successfully break MechanicalAnalysis.py into several files
