+++ src/Mod/Fem/FemTools.py
@@ -29,46 +29,8 @@ Significant refactoring(mainly renaming) work happends here, code logic is remai
 """ This is an adapter, before FemTools is fully moved to CaeSolver
 InspectionGui reference to this file, 
 """
+from CaeSolver import makeCaeSolver
 def FemTools():
     return makeCaeSolver('Calculix')
     
     
-#This list could be gen by python code by check subfolders
-registered_solvers={
-"Calculix":{ "name":"Calculix", "catogory":"Fem", "module": "ccxFemSolver","version": (2,7,0), "reader":"","writer":""}
-}
-
-from PreImport import *
-
-#ViewProvider and CaeSolver DocumentObject should be gen
-def makeCaeSolver(solverName, analysis=None):
-    """CaeSolver Factory method, solverName string  "Calculix" "OpenFoam"
-    """
-    #obj = FreeCAD.ActiveDocument.addObject("App::CaeSolverPython", name)  #todo
-    #solverName=FemGui.getActiveAnalysis().solverName  #ActiveAnalysis object is not constructed yet!
-    if solverName in registered_solvers.keys():
-        sinfo=registered_solvers[solverName]
-        
-        if solverName=='Calculix':
-            import ccxFemSolver
-            if analysis==None:
-                solver=ccxFemSolver.Solver(FemGui.getActiveAnalysis()) #Should get fom Fem, not from GUI
-            else:
-                solver=ccxFemSolver.Solver(analysis)
-            
-        #FemGui.getActiveAnalysis().Proxy.solver=solver  #CaeAnalysis instance has no attribute 
-        #print "makeCaeSolver() type of FemGui.getActiveAnalysis().Proxy.solver ", type(FemGui.getActiveAnalysis().Proxy.solver)
-        
-            FreeCAD.Console.PrintMessage("Solver {} is created\n".format(solverName))
-            
-        else:
-            raise Exception('NonGui mode makeCaeSolver() is not implemented')            
-        """
-        if FreeCAD.GuiUp:
-            #FreeCADGui.loadModule("FemGui")
-            FreeCADGui.loadModule(sinfo["module"])
-            FreeCADGui.doCommand("FemGui.getActiveAnalysis().Proxy.solver={}.Solver()"%sinfo["module"])
-            #ViewProviderCaeSolver(obj.ViewObject)
-        """
-        return solver
-
